Add transit curation and support building without transit

+++ otp/run/rename_and_run.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
-cp /data/${HEADWAY_AREA}.graph.obj /otp/graph.obj
-
-java ${JAVA_MEM_ARGS} -jar otp-shaded.jar --load /otp
+if [[ -s /data/${HEADWAY_AREA}.graph.obj ]]; then
+    cp /data/${HEADWAY_AREA}.graph.obj /otp/graph.obj
+    java ${JAVA_MEM_ARGS} -jar otp-shaded.jar --load /otp
+fi
